[#142] Failure to detect broken references
This commit modifies the JsonReference validation process by checking that the file the JsonReference URI is pointing at exists and is accessible. This is necessary to handle case sensitivity. The validation previously worked because validation was done on URIs and did not check that the file could be open. Hyperlinks to files with a different case than the one in the JsonReference previously failed but validation passed. Now, the JsonReference URI and the file name should have same case.
